Day Two - Transferring Data
Input connections take care of the actual data transfer.
- Can send data from city to pole to city.
- Can send data from city to pole to pole then city.

Need to test if it still works (it should!) when it's city to pole to
pole to another pole THEN to a city. The way it is set up it should work
no matter how many poles the connection extends from, it always has the
same source.